feat: CARITAS-286 * add new it tests and unit test to the AgencyController
