+++ src/test/java/de/caritas/cob/agencyservice/testHelper/PathConstants.java
@@ -13,7 +13,7 @@ public class PathConstants {
   public static final String PAGE_PARAM = "page";
   public static final String PER_PAGE_PARAM = "perPage";
   public static final String PATH_GET_LIST_OF_AGENCIES = "/agencies";
-  public static final String PATH_GET_LIST_OF_AGENCIES_PRIVATE = "/agencies/private";
+  public static final String PATH_GET_LIST_OF_AGENCIES_PRIVATE = "/agencies/by-tenant";
   public static final String PATH_GET_LIST_OF_AGENCIES_TOPICS = "/agencies/topics";
   public static final String PATH_GET_AGENCIES_WITH_IDS = "/agencies/";
   public static final String GET_AGENCY_PATH = ROOT_PATH + "/agencies";
